requirements.txt
flask
pandas
numpy
scikit-learn
matplotlib
plotly
scipy
tflite-runtime
Pillow
joblib
openpyxl
firebase-admin==6.0.1
gunicorn